Remove temporary the past events descriptions
Those were inconsistent and only available for the first few events.
Maybe we can get the rid of them to simplify updating the events list
with less efforts.
Also, the current cards component for the events haven't been designed
with those piece of content in mind.
Those could be brought back as soon as we give the design more :heart: