Upgrading to Ruby v3.2.2 and glitch fixing
Upgrading was required because bundling wasn't working locally + Ruby v2.7.3 is kinda old :eyes:

With the upgrade to latest ruby version came a few issues.

First, it required to add `net-ftp` to Gemfile, as it was blocking launch of middleman. I tried but could not really identify why this was needed, if somebody feels like exploring it and
giving feedback, I'd gladly hear it :smile:

Second, with latest middleman came an issue with date, it was kinda ignored until we use quotes around them (aka. make them strings that need to be later on translated into dates), as stated in this issue resolving comment [on middleman-blog issue #238](https://github.com/middleman/middleman-blog/issues/238#issuecomment-1173181864)